requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Cors==3.0.10
Flask-Reuploaded==1.2.0
gunicorn==20.1.0
pytest==7.0.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3
Flask-Migrate==3.1.0
Werkzeug==2.0.0
coverage==6.4.4
Flask-Bcrypt==1.0.1
flask_wtf  == 1.0.1
wtform == 1.0